Fix type error in voxelize_binary_mask method
Cast scalars array to float before assignment to binary_mask['mask']
to resolve mypy type incompatibility. The __setitem__ method expects
float arrays, but scalars could be uint8, int, or float64 depending
on the input values.

+++ pyvista/core/filters/data_set.py
@@ -7845,7 +7845,7 @@ class DataSetFilters(_BoundsSizeMixin, DataObjectFilters):
             if background_value == 0
             else np.ones(scalars_shape, dtype=scalars_dtype) * background_value
         )
-        binary_mask['mask'] = scalars
+        binary_mask['mask'] = scalars.astype(float)
         # Make sure that we have a clean triangle-strip polydata
         # Note: Poly was partially pre-processed earlier
         poly_ijk = poly_ijk.strip()
